Refacto : seperate graphql queries
